+++ packages/editor/src/Editor.vue
@@ -223,16 +223,26 @@ export default defineComponent({
   emits: ['props-panel-mounted', 'update:modelValue'],
 
   setup(props, { emit }) {
-    editorService.on('root-change', (value) => {
-      const node = editorService.get<MNode | null>('node');
-      const nodeId = node?.id || props.defaultSelected;
-      if (nodeId && node !== value) {
-        editorService.select(nodeId);
-      } else {
-        editorService.set('nodes', [value]);
+    editorService.on('root-change', (value, preValue) => {
+      const nodeId = editorService.get<MNode | null>('node')?.id || props.defaultSelected;
+      let node;
+      if (nodeId) {
+        node = editorService.getNodeById(nodeId);
       }
 
-      emit('update:modelValue', toRaw(editorService.get('root')));
+      if (node && node !== value) {
+        editorService.select(node.id);
+      } else if (value?.items?.length) {
+        editorService.select(value.items[0]);
+      } else if (value?.id) {
+        editorService.set('nodes', [value]);
+        editorService.set('parent', null);
+        editorService.set('page', null);
+      }
+
+      if (toRaw(value) !== toRaw(preValue)) {
+        emit('update:modelValue', value);
+      }
     });
 
     // 初始值变化，重新设置节点信息
@@ -306,13 +316,12 @@ export default defineComponent({
     );
 
     onUnmounted(() => {
-      editorService.destroy();
-      historyService.destroy();
-      propsService.destroy();
-      uiService.destroy();
-      componentListService.destroy();
-      storageService.destroy();
-      codeBlockService.destroy();
+      editorService.resetState();
+      historyService.resetState();
+      propsService.resetState();
+      uiService.resetState();
+      componentListService.resetState();
+      codeBlockService.resetState();
     });
 
     const services: Services = {