+++ packages/ui/src/index.ts
@@ -18,12 +18,12 @@
 
 import Button from './button';
 import Container from './container';
-import Empty from './empty';
 import Img from './img';
 import Overlay from './overlay';
 import Page from './page';
 import Qrcode from './qrcode';
 import Text from './text';
+
 const ui: Record<string, any> = {
   page: Page,
   container: Container,
@@ -32,7 +32,6 @@ const ui: Record<string, any> = {
   img: Img,
   qrcode: Qrcode,
   overlay: Overlay,
-  empty: Empty,
 };
 
 export default ui;